按照需求修改打代码 完善业务逻辑

+++ src/main/java/com/tjh/controller/UploadController.java
@@ -8,6 +8,7 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 
+import com.github.pagehelper.util.StringUtil;
 import com.tjh.pojo.Student;
 import com.tjh.service.StudentService;
 import com.tjh.util.ResultMessage;
@@ -33,16 +34,15 @@ public class UploadController {
         //得到原来文件名 如果存在 删除文件
         Student sqlStudent = this.studentService.loadOneStudentByStudentId(Integer.valueOf(studentId));
         String oldFile= sqlStudent.getStudentSource();
-        File oldFiles = new File(oldFile);
-        if(oldFiles.exists()) {
-            oldFiles.delete();
+        if (!StringUtil.isEmpty(oldFile)) {
+            File oldFiles = new File(oldFile);
+            if (oldFiles.exists()) {
+                oldFiles.delete();
+            }
         }
         Map<String,Object> map=new HashMap<>();
         map.put("code", 0);
         map.put("msg", "");
-
-
-
         // 1,得到老名字
         String oldName = mf.getOriginalFilename();
         // 2,得到tomcat里面的upload目录
@@ -51,7 +51,7 @@ public class UploadController {
         // 3,得到当前时间
         String currentDate = RandomStrUtils.getCurrentDateToStr();
         // 4,得到新的父目录的路径 并判断是否存在 如果不存在就创建
-        String newRealPath = realPath + "/" + currentDate;
+        String newRealPath = realPath + "\\" + currentDate;
         File parentFile = new File(newRealPath);
         if (!parentFile.exists()) {
             parentFile.mkdirs();// 创建文件夹
@@ -72,7 +72,7 @@ public class UploadController {
         map.put("data", data);
         Student student = new Student();
         student.setStudentId(Integer.valueOf(studentId));
-        student.setStudentSource(newRealPath+"/"+newName);
+        student.setStudentSource(newRealPath+"\\"+newName);
         this.studentService.updateStudentByStudentId(student);
         return map;
     }
@@ -81,7 +81,6 @@ public class UploadController {
     @ResponseBody
     public ResultMessage downLoadFile(Integer studentId, HttpServletRequest request, HttpServletResponse response) throws UnsupportedEncodingException {
         //判断空 建立文件模板
-
         response.setContentType("application/force-download");
         response.setCharacterEncoding("utf-8");
         //查询当前学生
@@ -109,8 +108,7 @@ public class UploadController {
             e.printStackTrace();
             return ResultMessage.success(ResultMessage.FAILCODE, ResultMessage.FAIL);
         }
-
-        return ResultMessage.success(ResultMessage.SUCCESSCODE, ResultMessage.SUCCESSFUL);
+             return ResultMessage.success(ResultMessage.SUCCESSCODE, ResultMessage.SUCCESSFUL);
     }
